обновит структуру папок для GamePage, добавит BoardPage, FinishPage, StartPage
