+++ src/routes/Game/routes/Finish/index.jsx
@@ -1,0 +1,7 @@
+const FinishPage = () => {
+  return (
+    <h1>This is page Finish</h1>
+  );
+};
+
+export default FinishPage;
